Updated OpenEnds and AddEnds()
Changed how OpenEnds are represented:
Was previously just a hash set of indices, now it is a dictionary of indices and accompanying OpenEnd struct object. This new OpenEnd struct contains a copy of the indices of the end, the number of connecting neighbors, and the path trace from the start of the area to the end.

Updated AddEnds():
Simplified some of the logic, and made AddEnds() solely responsible for documenting the path trace of a map cell. Also, this function now adds to BlockEntrances of a screen when it tries to place a secret end over an occupied cell.