Added ignore Directions to AddEnds
Added the ability to pass in a Directions, which will allow us to skip adding ends in those directions. This is used when calling AddEnds after placing the first and last levels in an area, and prevents errors from trying to place ends outside the map bounds or overwriting previously placed screens during CapEnds().